Change C++ client API to take split rows on table creation
This changes the existing create table split key API to take a
KuduPartialRow object instead of pre-encoded split keys. This has the
nice side effect of making it easier for clients to avoid specifying
invalid split keys, and removes the extra step of encoding the row key
manually. The Java API is changed to take a PartialRow, the KeyEncoder
class is removed, and PartialRows are constructable via a Schema
instance instead of a KuduTable instance.

This change is necessary for flexible partitioning so that split points
can be specified for partitioning schemes that have range partitions on
column sets other than the primary key. To avoid duplicating
partition key encoding logic to every client implementation (it is more
complex than primary key encoding), the master receives split points in
unencoded form, and performs the tablet split encoding itself.

Change-Id: I0c8a07cb310423bad55a65592c48e37acaeea361