KUDU-818: (2/many) flexible partitioning groundwork for C++ client & server
This commit lays the foundation for flexible partitioning in the C++ client and
server. The client and server are updated to track partition schemas along with
table schemas.

Remaining work until flexible partitioning is usable by clients:
  * Introduce partition pruning logic in Partition/PartitionSchema and take
    advantage of it in the client. Currently scan partitions are pruned by
    start/stop keys, which incidentally works because we allow only PK range
    partitions.
  * A corresponding commit to bring the Java client up to date, as well as
    partition pruning logic in Java.

This change breaks forwards and backwards compatibility with clients, but
maintains master and tablet server compatibility with on-disk data.

Change-Id: I7627113017298417fbd4dd565335bf6f2910aeca