OS X: fix lsof helper
This changes the command for getting the process tree on OS X to use 'pstree',
since the OS X version of 'ps' is quite limited. This adds an additional
'pstree' Homebrew package requirement for running the server processes on OS X.
We could figure out how pstree works on mac, but it's probably not worth the
effort.

Change-Id: I21bc5b4bb091385e16ef994819a9b0e751387de7