Add kudu::shared_ptr for public exported types
We currently expose std::tr1::shared_ptr in our public client interface, which
is going to make it difficult to update to c++11. In particular, (Clang's)
libc++ does not include any tr1 libraries. libc++ is required when compiling
c++11 applications on OS X.

This commit lays out the groundwork for a solution. It replaces uses of
std::tr1::shared_ptr in the public interface (in particular, shared_ptr's to the
KuduClient, KuduTable, and KuduSession types), with kudu::shared_ptr, which is
(currently) a typedef to std::tr1::shared_ptr. This maintains compatibility with
existing client applications, but it is recommended going forward that client
applications use kudu::shared_ptr when interacting with the Kudu API instead of
std::tr1::shared_ptr.

When we move to C++11, kudu::shared_ptr will be changed to be a typedef of
std::tr1::shared_ptr when linking against libstdc++ (as it is now), and a
typedef of std::shared_ptr when linking against libc++. Incidently, the standard
library which Kudu links to will be split along platform lines, with Linux
remaining on libstdc++, and OS X on libc++.

Users wishing to write applications which compile against Kudu on OS X and Linux
will need to use the `kudu::shared_ptr` type (or have some other way to ensure
they use the correct shared_ptr type), however if the client only needs their
application to compile on Linux, `std::tr1::shared_ptr` will suffice.

Change-Id: Ia9fe85281d89f90ef7a3949a030924d89c65e819