Add devtoolset to Clang gcc toolchain search path
Clang automatically searches for the latest GCC toolchain when compiling in
order to include the system libstdc++ headers and libraries when compiling. This
commit adds the standard Red Hat devtoolset location to that search process so
that the devtoolset toolchain will be automatically used if installed.

See https://bugzilla.redhat.com/show_bug.cgi?id=1139077 for more detail.

Change-Id: I4ee4f0dcd7ee56214645fac901cd0d4dd8544693