Enable Red Hat devtoolset in RHEL 6 Environments
We are moving to C++11 in the near future, and will require a C++11 compatible
compiler on all suported platforms. Since the RHEL 6 system compiler (GCC 4.4)
does not support C++11, we need an alternate. The Red Hat devtoolset provides a
toolchain with GCC 4.9. Libraries (such as libkudu.so) which are compiled with
the Red Hat devtoolset toolchain automatically statically link portions of the
standard library which are new for C++11. This will allow us to take advantage
of C++11 features in the client, but still remain compatible with stock RHEL 6
systems.

This commit modifies the Jenkins build and test scripts to take advantage of the
devtoolset toolchain, hides the std library symbols that the devtoolset
links into library objects, and adds the devtoolset location to the default gcc
toolchain search paths for the thirdparty LLVM compiler. Additionally, since a
system compiler which supports C++11 can now be assumed, the precompiled clang
compiler is no longer a dependency on Linux and has been removed.

Change-Id: I5505911044aa9fda8fb5010c6fe446c8f5c5616e

Build thirdparty clang with system compiler

We build LLVM 3.4.2 in thirdparty, but it requires a c++11 compiler to build.
Now that we require the system compiler to be c++11 compatible, we no longer
need the bootstrapping compiler.

Change-Id: I4800d9eec0a566bcf71c59abdeae6a6d8b6cb40a