Enable Red Hat devtoolset in RHEL 6 Environments
We are moving to C++11 in the near future, and will require a C++11 compatible
compiler on all suported platforms. Since the RHEL 6 system compiler (GCC 4.4)
does not support C++11, we need an alternate. The Red Hat devtoolset provides a
toolchain with GCC 4.9. Libraries (such as libkudu.so) which are compiled with
the Red Hat devtoolset toolchain automatically statically link portions of the
standard library which are new for C++11. This will allow us to take advantage
of C++11 features in the client, but still remain compatible with stock RHEL 6
systems.

This commit modifies the Jenkins build and test scripts to take advantage of the
devtoolset toolchain.

Change-Id: I5505911044aa9fda8fb5010c6fe446c8f5c5616e