KUDU-1165 (Part 1): improve scan optimization
This commit improves the scan optimizations done in the predicate_encoder.
Chiefly, implicit predicates which are specified as part of the upper and lower
scan bounds are lifted into proper predicates, so that they can be included in
the optimization pass. This will result in a more efficient scan with the
following type of query (for a table with integer key columns a and b):

SELECT * FROM my_table
WHERE b = 5
AND key >= (10, MIN_VALUE)
AND key < (11, MIN_VALUE)

Whereas before this would result in a scan over all values where a=10, now the
scan will be limited to rows where a=10 and b=5.

Additionally, moving all scan constraints into the simplified predicates will
make it easier to implement partition pruning.

Change-Id: I672e3545522a68362984991494fe90af648bb190