WIP: Add ColumnPredicate class
This predicate type is meant to replace the current ScanPredicate class.
It's designed to encapsulate any type of predicate clause that can be applied to
a column. Right now it has Range, and Equality variants, but in the future could
be expanded to included a Compound variant to handle IN or conjunctive clauses.

ColumnPredicate will be integrated into the client and tablet in a followup
commit.

Change-Id: I72cac35d7a168f96c2ee52afd284489e99755e3f