[java-client] implement KuduPredicate API
This commit adds a new class KuduPredicate, which corresponds to the new
ColumnPredicatePB predicate type for scans. The KuduPredicate allows specifying
exclusive upper bound and lower bound predicates on columns. This commit is a
first step towards predicate optimization and partition pruning, but stops short
of actually implementing those optimizations.

This commit does not attempt to change any callers of the now deprecated
ColumnRangePredicate API.

Change-Id: Icdca28139a2f4f15633cfd872e372429bad831cd