[c++-client]: minimal changes to support tables with non-covering range partitions
This commit introduces the minimum set of changes necessary to add support to
the C++ client for normal operations (open/write/scan) on tables with
non-covering range partitions. Tests are not included, since the ability to
create such tables has not been introduced in the master or client. These
changes have been tested by hand, and in tests included in a followup commit to
add create table with range bounds support. This minimal functionality is being
pushed earlier in order to allow the 0.9 client to be compatible with 0.10
servers which have non-covered range partitioned tables.

Change-Id: Ib25b7a57b14b3d1e4e6dca75b88dad7c19ba7565