Update Java client for new master GetTableLocations semantics
The Java client was relying on the master returning an empty tablet locations
list when the table was still in the process of being created. Now, the master
returns a specific error. This code fixes the table locations lookup code in
AsyncKuduClient for look for that specific error code, and retry the table
locations lookup automatically.

Change-Id: I80bf5661aed1ba435800211411b3273874e5bbcf