make election timeout jitter more aggressive
Existing election timeouts have very low variance, and is capped at a maximum
value. Having a variance cap is problematic because it could cause Raft to not
make progress when the RTT between nodes is greater than the cap.
Counter-intuitively, having a low variance in timeouts causes elections to take
longer since it leads to more frequent election retries. This commit removes the
cap and increases the variance.

Change-Id: I2c9dad820c2b7d4bc4b9e791b78222559cdf63c8