KUDU-1533 Spark Kudu Rdd/Dataframe upsert
This patch improves the Kudu SparkSQL integration in two ways:

1) Removed trait creatableRelationProvider from DefaultSource. This
is an improvement because this trait cannot be correctly implemented
for Kudu without support for table truncation and because some modes
require auto-table creation but there's no satisfactory mechanism to
specify things like the partition schema in that case.

2) Added {insert, update, upsert, delete}Rows methods to KuduContext.
This replaces and expands on the removed functionality of
DefaultSource.

The change is backwards-incompatible. Syntax like

df.write.options(newOptions).mode("append").kudu

no longer works, and now needs to be written like

kuduContext.insertRows(df, tableName)

Change-Id: Ib8e0d50fb74dc2ce5e757e8a56fc1e863f699822