[c++] ColumnPredicate simplification for inequalities on boundary values
Adds special casing for inequality predicates on minimum and maximum
values. This tightens partition pruning in boundary cases. New tests are
added to cover the specific transformations. We already have good test
coverage of boundary values in predicate-test.

These changes uncovered a bug in type traits where the minimum value for
float types was defined as the lowest discrete value instead of
-infinity, after adding the predicate simplification multiple test cases
in predicate-test failed.

Change-Id: I9bb150249a51a69fb85af67d34e85b7d7bed325c