[security] protect against master SASL negotiation short-circuit
In both clients we weren't checking that the local SASL client
considered the negotiation to be complete when the server sent us a
SASL_SUCCESS message. This would allow a malicious server to trick the
client into thinking it had authenticated the server, when in reality it
had not. This ended up being easier in the Java client because the JDK
SASL API includes an 'isComplete', and there is not equivalent in cyrus
SASL, so the state has to be tracked explicitly.

Change-Id: I8f3b3d4f47e887b48c1c704c900e9260c22cec3a