server_negotiation: improve error handling
Unlike in the client negotiator, the server negotiator should send an
error response back to the client on authentication failure.  We were
missing this in a few cases, notably around token and cert
authentication failures. This also adds a new token-specific rpc error
code which is used when the client should discard its token and request
a new one from the master.

Change-Id: I64f9c0f59aa608bf5078d65883a7d9a6fb186c04