kudu-mapreduce: tweak poms
This changes hadoop-client to be a provided dependency of
kudu-mapreduce, and switches kudu-mapreduce and kudu-client-tools from
using the assembly pluging to the shade plugin (like the rest of the
submodules). Normally changing a maven dependency from compile to
provided scope might be considered a breaking change, but because the
kudu-mapreduce artifact has previously not been published as a fat
jar[1], due to the configuration of the assembly plugin, the artifact
didn't include the hadoop-client classes (which I believe means there's
no binary compat risk). In the future user M/R applications (such as
kudu-client-tools) will need to explicitly add a dependency on
hadoop-client, if they use classes from it, but they should have been
doing this anyway. Going forward, kudu-mapreduce will be published as a
fat jar containing kudu-client. kudu-client-tools will be published as a
fat jar containing kudu-client and kudu-mapreduce.

[1]: https://mvnrepository.com/artifact/org.apache.kudu/kudu-mapreduce/1.2.0

Change-Id: I599ea9610fc3d541847b0205e959e94fb8425b2e