KUDU-2063 (part 3): Add support for databases to the CatalogManager
This reorganizes the internal CatalogManager state so that tables are
indexed by database-name -> table-name -> TableInfo, instead of
table-name -> TableInfo. I also took the opportunity to rename the
internal fields of CatalogManager.

As of this commit, databases are created as needed when a new table is
created. Follow up commits will add more layers: first class database
objects in the sys-catalog with their own IDs, and APIs for
creating/listing/dropping/renaming databases. I broke up the logic
explicitly into steps in order to call out that it is feasible for us to
have a lighter-weight database concept than a traditional RDBMS has.

No tests are added, since it's not yet possible to create databases
besides the default database (and that's covered by existing tests).

Change-Id: I02b3c7c162e19d7bd371b189a048fa73ea49bed6